Refactor twitter digest browser collector
