+++ twitter-digest/scripts/dom_scripts/extract_dm_conversation_text.js
@@ -1,0 +1,4 @@
+(() => {
+  const panel = document.querySelector('[data-testid="dm-conversation-panel"]') || document.querySelector('[data-testid="conversationPanel"]') || document.querySelector('main') || document.body;
+  return (panel.innerText || '').trim().slice(0, 12000);
+})()
