+++ twitter-digest/scripts/dom_scripts/extract_main_text.js
@@ -1,0 +1,4 @@
+(() => {
+  const main = document.querySelector('main') || document.body;
+  return (main.innerText || '').trim().slice(0, 12000);
+})()
