+++ twitter-digest/scripts/script_utils.py
@@ -2,6 +2,7 @@
 
 from __future__ import annotations
 
+import datetime as dt
 from pathlib import Path
 
 
@@ -18,3 +19,16 @@ def ensure_private_dir(path: Path) -> None:
         path.chmod(0o700)
     except PermissionError:
         pass
+
+
+def now_iso() -> str:
+    return dt.datetime.now().astimezone().isoformat()
+
+
+def now_stamp() -> str:
+    return dt.datetime.now().astimezone().strftime("%Y%m%d-%H%M%S")
+
+
+def local_timezone_name() -> str:
+    tz = dt.datetime.now().astimezone().tzinfo
+    return str(tz) if tz else "local"
